requirements/dev.txt
Django==5.2.5
django-db-logger==0.1.13
django-filter==25.1
djangorestframework==3.16.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.8.1
djangorestframework-simplejwt==5.5.1
django-rest-passwordreset==1.5.0
gunicorn==23.0.0
pytz==2025.2
psycopg2-binary==2.9.10
python-dotenv==1.1.1
logtail-python==0.3.4